NES.emu updates and misc tweaks/fixes
* Imagine: Use size_t for size type in VMemArray
* Imagine: Rename copy and transform algorithm function to allow easier switching between STL versions
* EmuFramework: Fix issue where input action table isn't correctly reset in InputDeviceData::buildKeyMap()
* EmuFramework: Add sound buffer size option of 1 frame
* 2600.emu: Directly set analog paddle position upon pointer input
* NES.emu: Update core to FCEUX 2.6.2
* NES.emu: Add mapper 319 support
* NES.emu: Add mid-frame audio updates

+++ imagine/include/imagine/util/container/VMemArray.hh
@@ -35,7 +35,7 @@ public:
 
 	constexpr VMemArray() = default;
 
-	VMemArray(unsigned size)
+	VMemArray(size_t size)
 	{
 		resize(size);
 	}
@@ -97,8 +97,8 @@ public:
 	T *data() { return data_; }
 	const T *data() const { return data_; }
 
-	T& operator[] (int idx) { return data()[idx]; }
-	const T& operator[] (int idx) const { return data()[idx]; }
+	T& operator[] (size_t idx) { return data()[idx]; }
+	const T& operator[] (size_t idx) const { return data()[idx]; }
 
 	void resetElements()
 	{
@@ -108,7 +108,7 @@ public:
 
 private:
 	T *data_{};
-	unsigned size_ = 0;
+	size_t size_ = 0;
 
 	void freeStorage()
 	{
